refactor: cli script divided into 4 smaller scripts
