+++ cli/extend-cradle.js
@@ -1,0 +1,40 @@
+import fs from "fs";
+import path from "path";
+
+export const extendCradle = (namePascal, nameCamel, nameKebab) => {
+    const diContainerPath = path.join(
+        process.cwd(),
+        "src/types/di-container.type.ts"
+    );
+
+    let diContent = fs.readFileSync(diContainerPath, "utf8");
+
+    if (!diContent.includes(`${namePascal}Service`)) {
+        const importMarker = 'import { EnvConfig } from "./env.type.js";';
+
+        const newImports = `
+import { ${namePascal}Service } from "@/modules/${nameCamel}/${nameKebab}.service.js";
+import { ${namePascal}Handler } from "@/modules/${nameCamel}/${nameKebab}.handler.js";`;
+
+        diContent = diContent.replace(
+            importMarker,
+            `${importMarker}${newImports}`
+        );
+    }
+
+    if (!diContent.includes(`${nameCamel}Service`)) {
+        const marker = "config: EnvConfig;";
+
+        const newCradleProps = `\n
+    ${nameCamel}Service: ${namePascal}Service;
+    ${nameCamel}Handler: ${namePascal}Handler;`;
+
+        diContent = diContent.replace(marker, `${marker}${newCradleProps}`);
+    }
+
+    fs.writeFileSync(diContainerPath, diContent);
+
+    console.log(
+        `✅ DI container updated with ${nameCamel} service and handler.`
+    );
+};
